(Qt) Combine qt download files into one - further binary size reduction
